+++ docs/en/changes/changes.md
@@ -65,6 +65,7 @@
 * BanyanDB: support group `replicas` and `user/password` for basic authentication.
 * BanyanDB: fix Zipkin query missing tag `QUERY`.
 * Fix `IllegalArgumentException: Incorrect number of labels`, tags in the `LogReportServiceHTTPHandler` and `LogReportServiceGrpcHandler` inconsistent with `LogHandler`.
+* BanyanDB: fix Zipkin query by `annotationQuery`
 
 #### UI